Stop recommending JNA as native terminal implementation
Jansi and JNA should be equivalent as far as functionality is
concerned, so choosing between them is a matter of preference.
TCA does not care which one is used.

Remove the explicit runtime dependency on the JNA terminal
implementation and document that an explicit runtime
dependency needs to be added for one of them.