Replace TCAHighlightConverter with TCALookup
Log4j has multiple converters that make use of ANSI colors. %highlight
is one of them, but there is also %style for example. We would need
to wrap them all to make them work correctly with TCA's ANSI detection.

It turns out there is actually another, easier approach that avoids that.
Log4j has another concept called "lookups" which are simply variables
that can be looked up in the configuration. Fortunately, they are also
easily extendable with Log4j's plugin architecture.

This allows to implement a ${tca:disableAnsi} property that can be hooked
up to the disableAnsi="..." configuration of PatternLayout. The advantage
is that there is no need to wrap any of Log4j's own internal objects and
it works for all the different patterns (including %highlight and %style).

It can be used in the configuration as follows:

  <PatternLayout pattern="..." disableAnsi="${tca:disableAnsi}">
  <PatternLayout>
      <LoggerNamePatternSelector defaultPattern="..." disableAnsi="${tca:disableAnsi}">